Add test mdLinks and option --help
