+++ package.json
@@ -15,9 +15,10 @@
     "test": "test"
   },
   "scripts": {
-    "test": "jest --coverage ",
+    "test": "jest --coverage  && npm run coverage",
     "testing": "jest --watch",
-    "md-links": "node src/index.js"
+    "md-links": "node src/index.js",
+    "coverage": "serve ./coverage/lcov-report"
   },
   "repository": {
     "type": "git",
@@ -41,7 +42,8 @@
     "eslint-config-airbnb-base": "15.0.0",
     "eslint-plugin-import": "2.26.0",
     "jest": "29.3.1",
-    "nodemon": "2.0.20"
+    "nodemon": "2.0.20",
+    "serve": "^14.1.2"
   },
   "dependencies": {
     "ascii-art": "2.8.5",
