--- pruebasPruebas/pruebas/otro.md
@@ -1,2 +1,0 @@
-[nodeJS](https://nodejs.org/api/path.html#pathextnamepath)
-[Curso nodeJS - midudev](https://www.youtube.com/watch?v=o85OkeVtm7k)
