--- pruebasPruebas/pruebas/otros.txt
@@ -1,1 +1,0 @@
-Solo es texto
