--- pruebasPruebas/pruebas/pruebaConLinks.md
@@ -1,8 +1,0 @@
-# Titulo 2
-
-# Con Links
-
-[nodeJS](https://nodejs.org/api/path.html#pathextnamepath)
-[Curso nodeJS - midudev](https://www.youtube.com/watch?v=o85OkeVtm7k)
-[Twitch](https://www.twitch.tv/midudev)
-
