--- pruebasPruebas/pruebas/pruebaSinLinks.md
@@ -1,5 +1,0 @@
-# Titulo
-
-## Algo
-
-Solo es una prueba
