+++ src/cli.js
@@ -3,7 +3,7 @@ const { message } = require('./api');
 
 const { log } = console;
 const { mdLinks } = require('./md_links');
-const { welcome } = require('./welcome');
+const { help } = require('./help');
 
 const resolveMDLinks = (path, options) => {
   mdLinks(path, options)
@@ -15,16 +15,18 @@ const resolveMDLinks = (path, options) => {
     });
 };
 
-welcome();
-
 const path = process.argv[2];
 const options = process.argv.slice(3);
 const option = options.join(' ');
-if (options.length === 0) {
+
+if (path === '--help' || path === 'h') {
+  help();
+} else if (options.length === 0) {
   resolveMDLinks(path, { validate: false });
 } else if (options.length >= 1) {
   switch (option) {
     case '--stats --validate':
+    case '--validate --stats':
       resolveMDLinks(path, { stats: true, validate: true });
       break;
     case '--validate':
@@ -32,6 +34,9 @@ if (options.length === 0) {
       break;
     case '--stats':
       resolveMDLinks(path, { stats: true });
+      break;
+    case '--help':
+      help();
       break;
     default:
       log(message('Opción no válida', 'red'));
